调整cloud相关代码

+++ takin-web-service-adapters/takin-web-adapter/takin-web-adapter-api/src/main/java/io/shulie/takin/adapter/api/model/request/resource/ResourceLockRequest.java
@@ -15,6 +15,6 @@ public class ResourceLockRequest extends ContextExt {
     private BigDecimal cpu;
     private BigDecimal memory;
     private Integer number;
-    private Integer watchmanId = WATCH_MAN_ID;
+    private Long watchmanId = WATCH_MAN_ID;
     private String callbackUrl;
 }
